fixed deletion of assets

+++ Admin/components/assetstuff.php
@@ -88,6 +88,15 @@
 					$message = "Success!";
 				}
 
+				
+			}
+		}
+	} else {
+		if(isset($_POST['type'])) {
+			if(isset($_POST['id'])) {
+				$asset = Asset::FromID(intval($_POST['id']));
+
+				if($asset != null && $asset->creator->id == $user->id || $user->IsAdmin())
 				if($_POST['type'] == "delete") {
 					$stmt = $con->prepare('DELETE FROM `inventory` WHERE `inv_assetid` = ?');
 					$stmt -> bind_param("i", $id);
@@ -97,17 +106,26 @@
 					$stmt -> bind_param("i", $id);
 					$stmt->execute();
 
-					$stmt = $con->prepare('DELETE FROM `assets` WHERE `asset_id` = ?');
+					$stmt = $con->prepare('DELETE FROM `visit` WHERE `visit_place` = ?');
+					$stmt -> bind_param("i", $id);
+					$stmt->execute();
+
+					$stmt = $con->prepare('DELETE FROM `favourites` WHERE `fav_assetid` = ?');
 					$stmt -> bind_param("i", $id);
 					$stmt->execute();
 					
 					CheckAndDeleteAsset($id);
-					$message = "Success!";
+
+					$stmt = $con->prepare('DELETE FROM `assets` WHERE `asset_id` = ?');
+					$stmt -> bind_param("i", $id);
+					$stmt->execute();
+					die("Success!");
 				}
 			}
 		}
-	} else {
+
 		$message = "You are not authorised to use this.";
+		
 	}
 
 	die($message);
